requirements.txt
discord.py
giphy_client
python-dotenv
